Implement the GP backend with botorch
