+++ pyproject.toml
@@ -64,7 +64,7 @@ checking = [
 ]
 document = [
     "cma",
-    "gpytorch",
+    "botorch",
     "distributed",
     "fvcore",
     "lightgbm",
@@ -114,7 +114,7 @@ integration = [
     "xgboost",
 ]
 optional = [
-    "gpytorch",  # optuna/terminator/gp/gpytorch
+    "botorch",  # optuna/terminator/gp/botorch
     "matplotlib!=3.6.0",  # optuna/visualization/matplotlib
     "pandas",  # optuna/study.py
     "plotly>=4.9.0",  # optuna/visualization.
